+++ web/src/utils/__tests__/locationNavigator.test.js
@@ -1,4 +1,4 @@
-import { navigateSpecifiedPteam } from "../../utils/locationNavigator";
+import { navigateSpecifiedPteam } from "../locationNavigator";
 
 test.each([
   // not navigate
